sentinelx_backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
